+++ src/app/(pages)/admin/role/history/page.tsx
@@ -10,6 +10,8 @@ import { useRouter } from "next/navigation";
 import Link from "next/link";
 import Button from "@/app/components/UI/Button";
 import { Role } from "@/app/model/Role";
+import PopConfirm from "@/app/components/UI/PopConfirm";
+import Toast from "@/app/components/UI/Toast";
 
 const columns: (keyof Role)[] = ["id", "role", "remark"];
 
@@ -17,10 +19,17 @@ const UserRoles = () => {
   const router = useRouter();
   const [roles, setRoles] = useState([]);
   const [loading, setLoading] = useState(true);
+  const [isPopConfirmOpen, setIsPopConfirmOpen] = useState(false);
+  const [roleToDelete, setRoleToDelete] = useState<number | null>(null);
+  const [toast, setToast] = useState<{
+    message: string;
+    type: "success" | "error";
+    position: "top-right";
+  } | null>(null);
   const [pagination, setPagination] = useState({
     totalDocs: 0,
     totalPages: 0,
-    page: 1, 
+    page: 1,
     limit: 10,
     pageCounter: 0,
     hasPrevPage: false,
@@ -32,7 +41,7 @@ const UserRoles = () => {
   }, []);
 
   const handlePageChange = (page: number) => {
-    setPagination(prevState => ({
+    setPagination((prevState) => ({
       ...prevState,
       page,
     }));
@@ -48,28 +57,41 @@ const UserRoles = () => {
         router.push(`/admin/role/update/${row.id}`);
         break;
       case "delete":
-        handleDelete(row.id);
+        setRoleToDelete(row.id);
+        setIsPopConfirmOpen(true);
         break;
       default:
         break;
     }
   };
 
-  const handleDelete = async (id: number) => {
+  const handleDelete = async () => {
+    if (roleToDelete === null) return;
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
-      const response = await fetch(`${apiUrl}/api/v1/roles/${id}`, {
+      const response = await fetch(`${apiUrl}/api/v1/roles/${roleToDelete}`, {
         method: "DELETE",
       });
 
       if (!response.ok) {
-        throw new Error("Failed to delete the region");
+        throw new Error("Failed to delete the Role");
       }
 
-      console.log(`Region with id ${id} deleted successfully`);
+      console.log(`Role with id ${roleToDelete} deleted successfully`);
       fetchRoles();
+      setToast({
+        message: "You have successfully deleted Role.",
+        type: "success",
+        position: "top-right",
+      });
+      setIsPopConfirmOpen(false);
     } catch (error) {
-      console.error("Error deleting the region: ", error);
+      console.error("Error deleting the Role: ", error);
+      setToast({
+        message: `${error.message}`,
+        type: "error",
+        position: "top-right",
+      });
     }
   };
 
@@ -78,11 +100,13 @@ const UserRoles = () => {
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
 
-      const response = await fetch(`${apiUrl}/api/v1/roles/all`, { method: "GET" });
+      const response = await fetch(`${apiUrl}/api/v1/roles/all`, {
+        method: "GET",
+      });
       if (response.ok) {
         const data = await response.json();
         setRoles(data.data);
-        setPagination(data.pagination)
+        setPagination(data.pagination);
         // setFilteredRegions(data);
       } else {
         console.error("Failed to fetch data");
@@ -94,51 +118,105 @@ const UserRoles = () => {
     }
   };
 
+  const searchRoles = async (searchQuery: any) => {
+    setLoading(true);
+    try {
+      const apiUrl = process.env.NEXT_PUBLIC_API_URL;
+      const response = await fetch(
+        `${apiUrl}/api/v1/roles/all?searchQuery=${searchQuery}&page=${pagination.page}&limit=${pagination.limit}`,
+        {
+          method: "GET",
+        }
+      );
+
+      if (response.ok) {
+        const data = await response.json();
+        setRoles(data.data);
+        setPagination(data.pagination);
+      } else {
+        console.error("Failed to fetch data");
+      }
+    } catch (error) {
+      console.error("Error fetching data:", error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleCancel = () => {
+    setIsPopConfirmOpen(false);
+    setRoleToDelete(null);
+  };
+
   const handleSearch = (query: string) => {
-    console.log("Searching for:", query);
+    searchRoles(query);
+  };
+
+  const handleClearSearch = () => {
+    searchRoles("");
+    fetchRoles();
   };
 
   return (
-    <BoxWrapper
-      icon={<FaExclamationTriangle />}
-      title="User Role"
-      borderColor="border-primary"
-      borderThickness="border-b-4"
-    >
-      <div className="flex flex-1 items-center justify-between mb-2 w-full">
-      <Search
-          onSearch={handleSearch}
-          placeholder="Search Role..."
-          buttonText="Search Role"
-        />
-        <Link href="/admin/role/create">
-          <Button
-            color="primary"
-            text="Create Role"
-            icon={<FaPlus />}
-            className="ml-auto"
-            size="large"
-            borderRadius={5}
+    <>
+      <BoxWrapper
+        icon={<FaExclamationTriangle />}
+        title="User Role"
+        borderColor="border-primary"
+        borderThickness="border-b-4"
+      >
+        <div className="flex flex-1 items-center justify-between mb-2 w-full">
+          <Search
+            onSearch={handleSearch}
+            onClear={handleClearSearch}
+            placeholder="Search Role..."
+            buttonText="Search Role"
           />
-        </Link>
-      </div>
-      {loading ? (
-        <div>Loading...</div>
-      ) : (
-        <>
-          <Table columns={columns} data={roles} onAction={handleAction} />
-          <div className="flex justify-end mt-4">
-            <Pagination
-              currentPage={pagination.page}
-              totalPages={pagination.totalPages}
-              onPageChange={handlePageChange}
-              hasNextPage={pagination.hasNextPage}
-              hasPrevPage={pagination.hasPrevPage}
+          <Link href="/admin/role/create">
+            <Button
+              color="primary"
+              text="Create Role"
+              icon={<FaPlus />}
+              className="ml-auto"
+              size="large"
+              borderRadius={5}
             />
-          </div>
-        </>
+          </Link>
+        </div>
+        {loading ? (
+          <div>Loading...</div>
+        ) : (
+          <>
+            <Table columns={columns} data={roles} onAction={handleAction} />
+            <div className="flex justify-end mt-4">
+              <Pagination
+                currentPage={pagination.page}
+                totalPages={pagination.totalPages}
+                onPageChange={handlePageChange}
+                hasNextPage={pagination.hasNextPage}
+                hasPrevPage={pagination.hasPrevPage}
+              />
+            </div>
+          </>
+        )}
+
+        <PopConfirm
+          isOpen={isPopConfirmOpen}
+          onConfirm={handleDelete}
+          onCancel={handleCancel}
+          message="Are you sure you want to delete this Role?"
+          title="Delete Role"
+        />
+      </BoxWrapper>
+      {toast && (
+        <Toast
+          message={toast.message}
+          type={toast.type}
+          position={toast.position}
+          onClose={() => setToast(null)}
+        />
       )}
-    </BoxWrapper>
+    </>
   );
 };
