+++ src/app/(pages)/admin/source/history/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useEffect, useState } from "react";
-import { FaExclamationTriangle, FaPlus } from "react-icons/fa";
+import { FaExclamationTriangle, FaInfoCircle, FaPlus } from "react-icons/fa";
 import BoxWrapper from "@/app/components/UI/BoxWrapper";
 import Table from "@/app/components/UI/Table";
 import Pagination from "@/app/components/UI/Pagination";
@@ -10,14 +10,22 @@ import { useRouter } from "next/navigation";
 import Link from "next/link";
 import Button from "@/app/components/UI/Button";
 import { Source } from "@/app/model/Source";
+import Toast from "@/app/components/UI/Toast";
+import PopConfirm from "@/app/components/UI/PopConfirm";
 
 const columns: (keyof Source)[] = ["id", "source", "remark"];
 
 const SourceOfInformation = () => {
   const router = useRouter();
-  const [currentPage, setCurrentPage] = useState(1);
   const [sources, setSources] = useState([]);
   const [loading, setLoading] = useState(true);
+  const [isPopConfirmOpen, setIsPopConfirmOpen] = useState(false);
+  const [sourceToDelete, setSourceToDelete] = useState<number | null>(null);
+  const [toast, setToast] = useState<{
+        message: string;
+        type: "success" | "error";
+        position: "top-right";
+      } | null>(null);
   const [pagination, setPagination] = useState({
     totalDocs: 0,
     totalPages: 0,
@@ -50,30 +58,45 @@ const SourceOfInformation = () => {
         router.push(`/admin/source/update/${row.id}`);
         break;
       case "delete":
-        handleDelete(row.id);
+        setSourceToDelete(row.id);
+        setIsPopConfirmOpen(true);
         break;
       default:
         break;
     }
   };
 
-  const handleDelete = async (id: number) => {
+  const handleDelete = async () => {
+    if (sourceToDelete === null) return;
+
     try {
       const apiUrl = process.env.NEXT_PUBLIC_API_URL;
-      const response = await fetch(`${apiUrl}/api/v1/sources/${id}`, {
+      const response = await fetch(`${apiUrl}/api/v1/sources/${sourceToDelete}`, {
         method: "DELETE",
       });
 
       if (!response.ok) {
-        throw new Error("Failed to delete the region");
+        throw new Error("Failed to delete the Source");
       }
-
-      console.log(`Region with id ${id} deleted successfully`);
+   
+      console.log(`Source with id ${sourceToDelete} deleted successfully`);
       fetchSources();
+      setToast({
+        message: "You have successfully deleted Source.",
+        type: "success",
+        position: "top-right",
+      });
+      setIsPopConfirmOpen(false); 
     } catch (error) {
-      console.error("Error deleting the region: ", error);
+      console.error("Error deleting the Source: ", error);
+      setToast({
+        message: `${error.message}`,
+        type: "error",
+        position: "top-right",
+      });
     }
   };
+
 
   const fetchSources = async () => {
     setLoading(true);
@@ -84,7 +107,26 @@ const SourceOfInformation = () => {
       if (response.ok) {
         const data = await response.json();
         setSources(data.data);
-        // setFilteredRegions(data);
+        setPagination(data.pagination)
+      } else {
+        console.error("Failed to fetch data");
+      }
+    } catch (error) {
+      console.error("Error fetching data:", error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const searchSources = async (searchQuery:string) => {
+    setLoading(true);
+    try {
+      const apiUrl = process.env.NEXT_PUBLIC_API_URL;
+
+      const response = await fetch(`${apiUrl}/api/v1/sources/all?searchQuery=${searchQuery}&page=${pagination.page}&limit=${pagination.limit}`, { method: "GET" });
+      if (response.ok) {
+        const data = await response.json();
+        setSources(data.data);
         setPagination(data.pagination)
       } else {
         console.error("Failed to fetch data");
@@ -97,12 +139,23 @@ const SourceOfInformation = () => {
   };
 
   const handleSearch = (query: string) => {
-    console.log("Searching for:", query);
+    searchSources(query)
+  };
+
+  const handleClearSearch = () =>{
+    searchSources("");  
+    fetchSources();
+  }
+
+  const handleCancel = () => {
+    setIsPopConfirmOpen(false);
+    setSourceToDelete(null);
   };
 
   return (
+    <>
     <BoxWrapper
-      icon={<FaExclamationTriangle />}
+      icon={<FaInfoCircle />}
       title="Source Of Information"
       borderColor="border-primary"
       borderThickness="border-b-4"
@@ -110,6 +163,7 @@ const SourceOfInformation = () => {
       <div className="flex flex-1 items-center justify-between m-2 w-full">
         <Search
           onSearch={handleSearch}
+          onClear={handleClearSearch}
           placeholder="Search Source..."
           buttonText="Search Source"
         />
@@ -143,7 +197,24 @@ const SourceOfInformation = () => {
           </div>
         </>
       )}
+
+      <PopConfirm
+        isOpen={isPopConfirmOpen}
+        onConfirm={handleDelete}
+        onCancel={handleCancel}
+        message="Are you sure you want to delete this Source?"
+        title="Delete Source"
+      />
     </BoxWrapper>
+        {toast && (
+          <Toast
+            message={toast.message}
+            type={toast.type}
+            position={toast.position}
+            onClose={() => setToast(null)}
+          />
+        )}
+        </>
   );
 };
